Made the create new superclass action fix the extends clauses.
The create new superclass action now updates the "extends" clause. Besides, if the selected class already extends another class, it makes the new class be the immediate superclass of the selected class and keeps the rest of the hierarchy intact.

Quick Assist now offers the create new superclass action only if the covering node is a type declaration.

This commit belongs to issue #1.